+++ lang/en/theme_boost_campus.php
@@ -108,19 +108,19 @@ $string['footerhidehomelinksetting'] = 'Hide link to the webpage start';
 
 // Additional layout setting.
 $string['additionallayoutsettings'] = 'Additional Layout Settings';
-// ...Badgearea.
-$string['badgeareaheadingsetting'] = 'Badgearea';
-$string['badgeareaheadingsetting_desc'] = 'The following settings allow adding an additional region for displaying images like badges or logos. This region will be added beneath the standard footer and above the optional footnote region. If images are uploaded this area will be displayed on all sites that use the columns2 layout.';
-$string['badgeareaitemssetting'] = 'Badge area items';
-$string['badgeareaitemssetting_desc'] = 'With this widget you can upload your images that will be displayed in the additional badge area region. The images will be sorted and displayed alphabetically by the filename. To remove this region, simply delete all uploaded images.';
-$string['badgeareaitemslinksetting'] = 'Badge area item links';
-$string['badgeareaitemslinksetting_desc'] = 'With this optional setting you can add links to your uplaoded images.<br/>
+// ...Image area.
+$string['imageareaheadingsetting'] = 'Image area';
+$string['imageareaheadingsetting_desc'] = 'The following settings allow adding an additional region for displaying images like logos. This region will be added beneath the standard footer and above the optional footnote region. If images are uploaded this area will be displayed on all sites that use the columns2 layout.';
+$string['imageareaitemssetting'] = 'Image area items';
+$string['imageareaitemssetting_desc'] = 'With this widget you can upload your images that will be displayed in the additional image area region. The images will be sorted and displayed alphabetically by the filename. To remove this region, simply delete all uploaded images.';
+$string['imageareaitemslinksetting'] = 'Image area item links';
+$string['imageareaitemslinksetting_desc'] = 'With this optional setting you can add links to your uplaoded images.<br/>
 Each line consists of the file identifier (the file name) the a link URL, separated by pipe characters. Each link declaration needs to be written in a new line. <br/>
 For example:<br/>
 moodle.jpg|http://moodle.org<br/>
-You can declare links for a abitrary amount of your uplaoded images. The links will be added only to those badges that match their filename with the identifier declared in this setting.';
-$string['badgeareaitemsmaxheightsetting'] = 'Badge area items maximal height';
-$string['badgeareaitemsmaxheightsetting_desc'] = 'With this setting you can change the height in pixels for your uploaded badges. All images will have the same maximum height and their width will be resized proportionally. The default value is set to 100 pixels.';
+You can declare links for a abitrary amount of your uplaoded images. The links will be added only to those images that match their filename with the identifier declared in this setting.';
+$string['imageareaitemsmaxheightsetting'] = 'Image area items maximal height';
+$string['imageareaitemsmaxheightsetting_desc'] = 'With this setting you can change the height in pixels for your uploaded images. All images will have the same maximum height and their width will be resized proportionally. The default value is set to 100 pixels.';
 // ...Footnote.
 $string['footnoteheadingsetting'] = 'Footnote';
 $string['footnoteheadingsetting_desc'] = 'The following setting allows to add an additional region for displaying a footnote.';
@@ -161,5 +161,5 @@ $string['darknavbarsetting'] = 'Dark navbar';
 $string['darknavbarsetting_desc'] = 'By checking this setting you can invert the default light navbar to a dark one with white links.';
 
 // ADDITIONAL STRINGS (IN ALPHABETICAL ORDER).
-$string['cachedef_badgearea'] = 'Cache for badgearea items';
+$string['cachedef_imagearea'] = 'Cache for imagearea items';
 $string['switchroleto'] = 'Switch role to';