+++ layout/login.php
@@ -38,7 +38,7 @@ echo $OUTPUT->render_from_template('theme_boost/login', $templatecontext);
 // MOFIFICATION START.
 // Include own layout file for the footnote region.
 // The theme_boost/login template already renders the standard footer.
-// The footer blocks and the badge area are currently not shown on the login page.
+// The footer blocks and the image area are currently not shown on the login page.
 // Here, we will add the footnote only.
 // Get footnote config.
 $footnote = get_config('theme_boost_campus', 'footnote');